restructured classes into separate files
