+++ google/cloud/bigtable/data/_metrics/__init__.py
@@ -1,0 +1,27 @@
+# Copyright 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from google.cloud.bigtable.data._metrics.handlers.opentelemetry import OpenTelemetryMetricsHandler
+from google.cloud.bigtable.data._metrics.handlers.stdout import StdoutMetricsHandler
+from google.cloud.bigtable.data._metrics.metrics_controller import BigtableClientSideMetricsController
+
+from google.cloud.bigtable.data._metrics.data_model import OperationType
+from google.cloud.bigtable.data._metrics.data_model import ActiveOperationMetric
+
+__all__ = (
+    "BigtableClientSideMetricsController",
+    "OpenTelemetryMetricsHandler",
+    "StdoutMetricsHandler",
+    "OperationType",
+    "ActiveOperationMetric",
+)
